+++ lib/micha_core.dart
@@ -2,11 +2,13 @@ library micha_core;
 
 export 'src/async/async_builder.dart';
 export 'src/async/spinner.dart';
+export 'src/extensions/ansi_formatting.dart';
 export 'src/extensions/by_name_or_null.dart';
 export 'src/extensions/separated.dart';
 export 'src/extensions/times.dart';
 export 'src/extensions/wrapped.dart';
 export 'src/layout/gap.dart';
+export 'src/logging.dart';
 export 'src/pagination/controls/current_page_label.dart';
 export 'src/pagination/controls/jump_to_next_page.dart';
 export 'src/pagination/controls/jump_to_numbered_page.dart';